requirements.txt
tkinterdnd2==0.3.0
pytesseract==0.3.10
pdf2image==1.16.3
pillow>=10.3.0  # Versão que suporta Python 3.13
pandas==2.2.1
openpyxl==3.1.2
pypdf2==3.0.1
pdfplumber==0.10.3
pyautogui==0.9.54
python-docx==1.1.0
numpy==1.26.4